pyproject.toml
[project]
name = "rag-powered-financial-product-explainer-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.0",
    "langchain-cerebras>=0.8.2",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.2.0",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.0.4",
    "livekit-agents[aws,deepgram,langchain,silero]>=1.3.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
]
